requirements.txt
torch>=1.9.0
pandas>=1.3.0
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0
mlflow>=1.20.0
pyyaml>=5.4.0
asteval>=0.9.0
psycopg2-binary>=2.9.0
minio>=7.0.0 